refactoring keep_up logic
