+++ node/src/reactor/main_reactor/control.rs
@@ -1,36 +1,31 @@
 use std::time::Duration;
-use tracing::{debug, error, info, trace, warn};
+use tracing::{debug, info, trace};
 
 use casper_hashing::Digest;
 use casper_types::{EraId, PublicKey, Timestamp};
 
 use crate::{
     components::{
-        block_accumulator::{StartingWith, SyncInstruction},
         block_synchronizer,
         block_synchronizer::BlockSynchronizerProgress,
         consensus::EraReport,
         contract_runtime::ExecutionPreState,
         deploy_buffer::{self, DeployBuffer},
-        diagnostics_port, event_stream_server, network, rest_server, rpc_server, sync_leaper,
-        sync_leaper::LeapStatus,
-        upgrade_watcher, InitializedComponent,
+        diagnostics_port, event_stream_server, network, rest_server, rpc_server, upgrade_watcher,
+        InitializedComponent,
     },
-    effect::{
-        announcements::ControlAnnouncement, requests::BlockSynchronizerRequest, EffectBuilder,
-        EffectExt, Effects,
-    },
+    effect::{announcements::ControlAnnouncement, EffectBuilder, EffectExt, Effects},
     fatal,
     reactor::{
         self,
         main_reactor::{
-            catch_up::CatchUpInstruction, keep_up_instruction::KeepUpInstruction,
+            catch_up::CatchUpInstruction, keep_up::KeepUpInstruction,
             upgrade_shutdown_instruction::UpgradeShutdownInstruction,
             upgrading_instruction::UpgradingInstruction, utils,
             validate_instruction::ValidateInstruction, MainEvent, MainReactor, ReactorState,
         },
     },
-    types::{BlockHash, BlockPayload, FinalizedBlock, Item, SyncLeapIdentifier},
+    types::{BlockHash, BlockPayload, FinalizedBlock},
     NodeRng,
 };
 
@@ -41,6 +36,113 @@ const VALIDATION_STATUS_DELAY_FOR_NON_SWITCH_BLOCK: Duration = Duration::from_se
 const DELAY_BEFORE_SHUTDOWN: Duration = Duration::from_secs(2);
 
 impl MainReactor {
+    pub(super) fn should_shutdown_for_upgrade(&self) -> bool {
+        if let Some(block_header) = self.recent_switch_block_headers.last() {
+            return self
+                .upgrade_watcher
+                .should_upgrade_after(block_header.era_id());
+        }
+        false
+    }
+
+    pub(super) fn should_commit_upgrade(&self) -> Result<bool, String> {
+        let highest_switch_block_header = match &self.switch_block {
+            None => {
+                return Ok(false);
+            }
+            Some(header) => header,
+        };
+
+        if !self
+            .chainspec
+            .protocol_config
+            .is_last_block_before_activation(highest_switch_block_header)
+        {
+            return Ok(false);
+        }
+
+        match self
+            .chainspec
+            .is_in_modified_validator_set(self.consensus.public_key())
+        {
+            None => match highest_switch_block_header.next_era_validator_weights() {
+                None => Err("switch_block should have next era validator weights".to_string()),
+                Some(next_era_validator_weights) => {
+                    Ok(next_era_validator_weights.contains_key(self.consensus.public_key()))
+                }
+            },
+            Some(is_validator) => Ok(is_validator),
+        }
+    }
+
+    pub(super) fn should_validate(
+        &mut self,
+        effect_builder: EffectBuilder<MainEvent>,
+        rng: &mut NodeRng,
+    ) -> (bool, Option<Effects<MainEvent>>) {
+        // if on a switch block, check if we should go to Validate
+        if self.switch_block.is_some() {
+            match self.create_required_eras(effect_builder, rng) {
+                Err(msg) => {
+                    return (false, Some(fatal!(effect_builder, "{}", msg).ignore()));
+                }
+                Ok(Some(effects)) => {
+                    return (true, Some(effects));
+                }
+                Ok(None) => (),
+            }
+        }
+        (false, None)
+    }
+
+    pub(super) fn refresh_contract_runtime(
+        &mut self,
+        next_block_height: u64,
+        pre_state_root_hash: Digest,
+        parent_hash: BlockHash,
+        parent_seed: Digest,
+    ) -> Result<(), String> {
+        let initial_pre_state = ExecutionPreState::new(
+            next_block_height,
+            pre_state_root_hash,
+            parent_hash,
+            parent_seed,
+        );
+        self.contract_runtime
+            .set_initial_state(initial_pre_state)
+            .map_err(|err| err.to_string())?;
+        Ok(())
+    }
+
+    pub(super) fn update_last_progress(
+        &mut self,
+        block_synchronizer_progress: &BlockSynchronizerProgress,
+        is_sync_back: bool,
+    ) {
+        if let BlockSynchronizerProgress::Syncing(_, _, last_progress) = block_synchronizer_progress
+        {
+            // do idleness / reattempt checking
+            let sync_progress = *last_progress;
+            if sync_progress > self.last_progress {
+                self.last_progress = sync_progress;
+                // if any progress has been made, reset attempts
+                self.attempts = 0;
+                let state = if is_sync_back {
+                    "Historical".to_string()
+                } else {
+                    format!("{}", self.state)
+                };
+                debug!(
+                    "{}: last_progress: {} {}",
+                    state, self.last_progress, block_synchronizer_progress
+                );
+            }
+            if self.last_progress.elapsed() > self.idle_tolerance {
+                self.attempts += 1;
+            }
+        }
+    }
+
     pub(super) fn crank(
         &mut self,
         effect_builder: EffectBuilder<MainEvent>,
@@ -367,234 +469,6 @@ impl MainReactor {
         }
     }
 
-    fn keep_up_instruction(
-        &mut self,
-        effect_builder: EffectBuilder<MainEvent>,
-        rng: &mut NodeRng,
-    ) -> KeepUpInstruction {
-        if self.should_shutdown_for_upgrade() {
-            return KeepUpInstruction::ShutdownForUpgrade;
-        }
-
-        match self.should_validate(effect_builder, rng) {
-            (true, Some(effects)) => {
-                info!("KeepUp: go to Validate");
-                return KeepUpInstruction::Validate(effects);
-            }
-            (_, Some(effects)) => {
-                // shutting down per consensus
-                return KeepUpInstruction::Do(Duration::ZERO, effects);
-            }
-            (_, None) => {
-                // remain in KeepUp
-            }
-        }
-        let forward_progress = self.block_synchronizer.forward_progress();
-        self.update_last_progress(&forward_progress, false);
-        let starting_with = match forward_progress {
-            BlockSynchronizerProgress::Idle => match self.storage.read_highest_complete_block() {
-                Ok(Some(block)) => {
-                    let block_height = block.height();
-                    let state_root_hash = block.state_root_hash();
-                    let block_hash = block.id();
-                    let accumulated_seed = block.header().accumulated_seed();
-                    match self.refresh_contract_runtime(
-                        block_height + 1,
-                        *state_root_hash,
-                        block_hash,
-                        accumulated_seed,
-                    ) {
-                        Ok(_) => StartingWith::LocalTip(
-                            block.id(),
-                            block_height,
-                            block.header().era_id(),
-                        ),
-                        Err(msg) => {
-                            return KeepUpInstruction::Fatal(msg);
-                        }
-                    }
-                }
-                Ok(None) => {
-                    error!("KeepUp: block synchronizer idle, local storage has no complete blocks");
-                    self.block_synchronizer.purge();
-                    return KeepUpInstruction::CatchUp;
-                }
-                Err(error) => {
-                    return KeepUpInstruction::Fatal(format!(
-                        "failed to read highest complete block: {}",
-                        error
-                    ))
-                }
-            },
-            BlockSynchronizerProgress::Syncing(block_hash, block_height, _) => match block_height {
-                None => StartingWith::Hash(block_hash),
-                Some(height) => StartingWith::BlockIdentifier(block_hash, height),
-            },
-            BlockSynchronizerProgress::Synced(block_hash, block_height, era_id) => {
-                debug!("KeepUp: synced block: {}", block_hash);
-                self.block_synchronizer.purge_forward();
-                StartingWith::SyncedBlockIdentifier(block_hash, block_height, era_id)
-            }
-        };
-        debug!(
-            ?starting_with,
-            "KeepUp: starting with {}",
-            starting_with.block_hash()
-        );
-        let sync_instruction = self.block_accumulator.sync_instruction(starting_with);
-        debug!(
-            ?sync_instruction,
-            "KeepUp: sync_instruction {}",
-            sync_instruction.block_hash()
-        );
-        match sync_instruction {
-            SyncInstruction::Leap { .. } => {
-                self.block_synchronizer.purge();
-                return KeepUpInstruction::CatchUp;
-            }
-            SyncInstruction::BlockSync { block_hash } => {
-                debug!("KeepUp: BlockSync: {:?}", block_hash);
-                if self.block_synchronizer.register_block_by_hash(
-                    block_hash,
-                    false,
-                    true,
-                    self.chainspec.core_config.simultaneous_peer_requests,
-                ) {
-                    info!(
-                        ?block_hash,
-                        "KeepUp: BlockSync: registered block by hash {}", block_hash
-                    );
-                    return KeepUpInstruction::Do(
-                        Duration::ZERO,
-                        effect_builder.immediately().event(|_| {
-                            MainEvent::BlockSynchronizerRequest(BlockSynchronizerRequest::NeedNext)
-                        }),
-                    );
-                }
-            }
-            SyncInstruction::CaughtUp { .. } => {
-                // noop
-            }
-        }
-        if false == self.sync_to_historical {
-            // if nothing else needs to be done, check later
-            return KeepUpInstruction::CheckLater(
-                "at perceived tip of chain".to_string(),
-                self.control_logic_default_delay.into(),
-            );
-        }
-        let historical_progress = self.block_synchronizer.historical_progress();
-        self.update_last_progress(&historical_progress, true);
-        match self.maybe_parent_block_identifier(&historical_progress) {
-            Err(msg) => KeepUpInstruction::Fatal(msg),
-            Ok(None) => KeepUpInstruction::CheckLater(
-                format!("Historical: syncing {:?}", historical_progress),
-                self.control_logic_default_delay.into(),
-            ),
-            Ok(Some((parent_hash, era_id))) => {
-                if false == self.validator_matrix.has_era(&era_id) {
-                    debug!("Historical: sync leaping for: {}", parent_hash);
-                    let leap_status = self.sync_leaper.leap_status();
-                    debug!("Historical: {:?}", leap_status);
-                    match leap_status {
-                        ls @ LeapStatus::Idle | ls @ LeapStatus::Failed { .. } => {
-                            if let LeapStatus::Failed {
-                                error,
-                                sync_leap_identifier: _,
-                                from_peers: _,
-                                in_flight: _,
-                            } = ls
-                            {
-                                self.attempts += 1;
-                                if self.attempts > self.max_attempts {
-                                    // self.crank will ensure shut down if no other progress
-                                    // is made before this event is processed
-                                    let msg = format!(
-                                        "Historical: failed leap back exceeded reattempt tolerance: {}",
-                                        error,
-                                    );
-                                    warn!("{}", msg);
-                                    return KeepUpInstruction::CheckLater(msg, Duration::ZERO);
-                                }
-                                error!("Historical: sync leap failed: {:?}", error);
-                            }
-                            let peers_to_ask = self.net.fully_connected_peers_random(
-                                rng,
-                                self.chainspec.core_config.simultaneous_peer_requests as usize,
-                            );
-                            let sync_leap_identifier =
-                                SyncLeapIdentifier::sync_to_historical(parent_hash);
-                            let effects = effect_builder.immediately().event(move |_| {
-                                MainEvent::SyncLeaper(sync_leaper::Event::AttemptLeap {
-                                    sync_leap_identifier,
-                                    peers_to_ask,
-                                })
-                            });
-                            info!("Historical: initiating sync leap for: {}", parent_hash);
-                            KeepUpInstruction::Do(Duration::ZERO, effects)
-                        }
-                        LeapStatus::Received {
-                            best_available,
-                            from_peers: _,
-                            ..
-                        } => {
-                            info!(
-                                "Historical: sync leap received for: {:?}",
-                                best_available.trusted_block_header.block_hash()
-                            );
-                            let era_validator_weights = best_available.era_validator_weights(
-                                self.validator_matrix.fault_tolerance_threshold(),
-                            );
-                            for evw in era_validator_weights {
-                                let era_id = evw.era_id();
-                                if self.validator_matrix.register_era_validator_weights(evw) {
-                                    info!("Historical: got era: {}", era_id);
-                                } else {
-                                    debug!("Historical: already had era: {}", era_id);
-                                }
-                            }
-                            KeepUpInstruction::CheckLater(
-                                "Historical: sync leap received".to_string(),
-                                Duration::ZERO,
-                            )
-                        }
-                        LeapStatus::Awaiting { .. } => KeepUpInstruction::CheckLater(
-                            "Historical: sync leap awaiting".to_string(),
-                            self.control_logic_default_delay.into(),
-                        ),
-                    }
-                } else if self.block_synchronizer.register_block_by_hash(
-                    parent_hash,
-                    true,
-                    true,
-                    self.chainspec.core_config.simultaneous_peer_requests,
-                ) {
-                    info!("Historical: register_block_by_hash: {}", parent_hash);
-                    let peers_to_ask = self.net.fully_connected_peers_random(
-                        rng,
-                        self.chainspec.core_config.simultaneous_peer_requests as usize,
-                    );
-                    debug!("Historical: peers count: {:?}", peers_to_ask.len());
-                    self.block_synchronizer
-                        .register_peers(parent_hash, peers_to_ask);
-
-                    return KeepUpInstruction::Do(
-                        Duration::ZERO,
-                        effect_builder.immediately().event(|_| {
-                            MainEvent::BlockSynchronizerRequest(BlockSynchronizerRequest::NeedNext)
-                        }),
-                    );
-                } else {
-                    self.block_synchronizer.purge_historical();
-                    KeepUpInstruction::CheckLater(
-                        "Historical: purged".to_string(),
-                        self.control_logic_default_delay.into(),
-                    )
-                }
-            }
-        }
-    }
-
     fn validate_instruction(
         &mut self,
         effect_builder: EffectBuilder<MainEvent>,
@@ -627,90 +501,6 @@ impl MainReactor {
             Ok(None) => ValidateInstruction::KeepUp,
             Err(msg) => ValidateInstruction::Fatal(msg),
         }
-    }
-
-    pub(super) fn update_last_progress(
-        &mut self,
-        block_synchronizer_progress: &BlockSynchronizerProgress,
-        is_sync_back: bool,
-    ) {
-        if let BlockSynchronizerProgress::Syncing(_, _, last_progress) = block_synchronizer_progress
-        {
-            // do idleness / reattempt checking
-            let sync_progress = *last_progress;
-            if sync_progress > self.last_progress {
-                self.last_progress = sync_progress;
-                // if any progress has been made, reset attempts
-                self.attempts = 0;
-                let state = if is_sync_back {
-                    "Historical".to_string()
-                } else {
-                    format!("{}", self.state)
-                };
-                debug!(
-                    "{}: last_progress: {} {}",
-                    state, self.last_progress, block_synchronizer_progress
-                );
-            }
-            if self.last_progress.elapsed() > self.idle_tolerance {
-                self.attempts += 1;
-            }
-        }
-    }
-
-    fn maybe_parent_block_identifier(
-        &mut self,
-        block_synchronizer_progress: &BlockSynchronizerProgress,
-    ) -> Result<Option<(BlockHash, EraId)>, String> {
-        if matches!(
-            block_synchronizer_progress,
-            BlockSynchronizerProgress::Syncing(_, _, _)
-        ) {
-            return Ok(None);
-        }
-        if let Some(block_header) = self.storage.get_highest_orphaned_block_header() {
-            if block_header.is_genesis() {
-                self.block_synchronizer.purge_historical();
-                return Ok(None);
-            }
-            debug!(
-                "Historical: attempting({}) for: {}",
-                block_header.height().saturating_sub(1),
-                block_header.parent_hash()
-            );
-            match self.storage.read_block_header(block_header.parent_hash()) {
-                Ok(Some(parent)) => Ok(Some((parent.block_hash(), parent.era_id()))),
-                Ok(None) => match block_header.era_id().predecessor() {
-                    Some(previous_era_id) => {
-                        Ok(Some((*block_header.parent_hash(), previous_era_id)))
-                    }
-                    None => Ok(None),
-                },
-                Err(err) => Err(err.to_string()),
-            }
-        } else {
-            Ok(None)
-        }
-    }
-
-    fn should_validate(
-        &mut self,
-        effect_builder: EffectBuilder<MainEvent>,
-        rng: &mut NodeRng,
-    ) -> (bool, Option<Effects<MainEvent>>) {
-        // if on a switch block, check if we should go to Validate
-        if self.switch_block.is_some() {
-            match self.create_required_eras(effect_builder, rng) {
-                Err(msg) => {
-                    return (false, Some(fatal!(effect_builder, "{}", msg).ignore()));
-                }
-                Ok(Some(effects)) => {
-                    return (true, Some(effects));
-                }
-                Ok(None) => (),
-            }
-        }
-        (false, None)
     }
 
     fn create_required_eras(
@@ -835,36 +625,6 @@ impl MainReactor {
             .ignore())
     }
 
-    pub(super) fn should_commit_upgrade(&self) -> Result<bool, String> {
-        let highest_switch_block_header = match &self.switch_block {
-            None => {
-                return Ok(false);
-            }
-            Some(header) => header,
-        };
-
-        if !self
-            .chainspec
-            .protocol_config
-            .is_last_block_before_activation(highest_switch_block_header)
-        {
-            return Ok(false);
-        }
-
-        match self
-            .chainspec
-            .is_in_modified_validator_set(self.consensus.public_key())
-        {
-            None => match highest_switch_block_header.next_era_validator_weights() {
-                None => Err("switch_block should have next era validator weights".to_string()),
-                Some(next_era_validator_weights) => {
-                    Ok(next_era_validator_weights.contains_key(self.consensus.public_key()))
-                }
-            },
-            Some(is_validator) => Ok(is_validator),
-        }
-    }
-
     fn commit_upgrade(
         &mut self,
         effect_builder: EffectBuilder<MainEvent>,
@@ -916,33 +676,5 @@ impl MainReactor {
             },
             Err(msg) => Err(msg),
         }
-    }
-
-    pub(super) fn should_shutdown_for_upgrade(&self) -> bool {
-        if let Some(block_header) = self.recent_switch_block_headers.last() {
-            return self
-                .upgrade_watcher
-                .should_upgrade_after(block_header.era_id());
-        }
-        false
-    }
-
-    fn refresh_contract_runtime(
-        &mut self,
-        next_block_height: u64,
-        pre_state_root_hash: Digest,
-        parent_hash: BlockHash,
-        parent_seed: Digest,
-    ) -> Result<(), String> {
-        let initial_pre_state = ExecutionPreState::new(
-            next_block_height,
-            pre_state_root_hash,
-            parent_hash,
-            parent_seed,
-        );
-        self.contract_runtime
-            .set_initial_state(initial_pre_state)
-            .map_err(|err| err.to_string())?;
-        Ok(())
     }
 }
