--- node/src/reactor/main_reactor/keep_up_instruction.rs
@@ -1,12 +1,0 @@
-use std::time::Duration;
-
-use crate::{effect::Effects, reactor::main_reactor::MainEvent};
-
-pub(super) enum KeepUpInstruction {
-    Validate(Effects<MainEvent>),
-    Do(Duration, Effects<MainEvent>),
-    CheckLater(String, Duration),
-    CatchUp,
-    ShutdownForUpgrade,
-    Fatal(String),
-}
